refactor: grande reorganisation pour faciliter linkage
